VoiceToText/requirements.txt
openai-whisper
torch
torchaudio
tkinterdnd2
mutagen